Dynamic initialization of data array
Currently we initialize the data array by explicitly setting its values
one by one. That obviously won't scale once we increase the number of
allowed items. At the same time, we cannot use the construct of
[value; N], since RefCell does not implement Copy trait.

In order to overcome this issues, this patch introduces use of
MaybeUninit to initialize an empty data array and only later populate it
with dynamically generated values.

Currently we initialize all the items of the array in the constructor of
Rack. That could get quite costly for larger sizes, so we may want to
handle the initialization lazily in a follow-up patch.

Signed-off-by: Petr Horacek <phoracek@redhat.com>