Populate data only on when `add` is called
The current code requires us to specify the default value of the stored
type so we can populate all data with it during Rack initialization.
For i32 we simply use 0 as the initial value, but that would become more
problematic with user-defined structure types.

With this patch, we keep the content of data uninitialized using
MaybeUninit and initalize it only once Rack::add is called with the
desired value.

Signed-off-by: Petr Horacek <phoracek@redhat.com>