Initialize Rack's data explicitly
The current method of Rack's data initialization using MaybeUninit block
has one issue, it cannot be applied on a generic type due to sizing
issues. Therefore, this patch replaces it with an explicit
initialization. This should be eventually replaced by the "Constants in
array repeat expressions" once it is implemented in Rust.

Signed-off-by: Petr Horacek <phoracek@redhat.com>