Return error when Rack is full
Currently we only raise a panic when there is no more space left on the
Rack, but it makes sense to return an Error and let the user decide how
to tackle the issue.

This patch changes the add method to return an Error when the Rack is
full and introduces a new must_add method that panics in case there is
no more space.

Signed-off-by: Petr Horacek <phoracek@redhat.com>